// src/types/enums.ts
export enum SortingOptions {
    PRICE = 'PRICE',
    DESC_DATE = 'DESC_DATE',
    ASC_DATE = 'ASC_DATE'
}

export enum DateSortingType {
    ASC = 'ASC',
    DESC = 'DESC'
}

export enum UserState {
    TRIAL = 'TRIAL',
    TRIAL_END = 'TRIAL_END',
    BLOCKED = 'BLOCKED',
    SUB_ACTIVE = 'SUB_ACTIVE'
}

export enum UserType {
    GOOGLE = 'GOOGLE',
    FACEBOOK = 'FACEBOOK',
    STANDARD = 'STANDARD',
}

export enum AccountStatus {
    NOT_VERIFIED = 'NOT_VERIFIED',
    VERIFIED = 'VERIFIED',
    EXPIRED = 'EXPIRED',
    TRIAL = 'TRIAL'
}

export enum BundleType {
    MONTHLY = 'MONTHLY',
    ONE_TIME = 'ONE_TIME'
}